Setting the userconfig by command
The users can now set their preferences by issueing the '/set' command. This is done in the following way:

/set INFB true

The command will check, whether INFB is a valid caourse and whether true is valid boolean (false is the alternative *smart*). These settings will be stored in the database. This is implemented as an early implementation for the usersettings, which ultimately will be implemented by using inline keyboard functionalities.