Improves overall method structure to support new URL calls
The previous version called the different URLs in order to receive the news for each course. The new approach calls one API URL and receives all the information from there. The methods had to be changed in order to work with this appraoch, as mapping the data correctly and so on changed. This seems to work now.