Adds rolling files for logging
This is done to avoid overly large logging files. Logs will from now on be stored in the ./logs folder from where the bot is run. The current log remain ```bot.log```, but its contents are moved over to the daily log file within the ./logs/ directory.
In addition to the daily logs, logs that are older than 90 days or if log files get tooo bloated (> 10GB of overall log file size), they will get removed.