Fixes the perception of old faculty messages being new
The error was caused, by an updated publicationDate, if old faculty news entries were updated. This resulted in a different hash for what was an old entry and resulted in old messages beeing pushed as new messages.
By creating a hash from the title and the date (not the publicationDate), this error can be avoided. Updated messages are ignored.

The error cannot be avoided, since the facultyNews entries don't have a unique id.