Enables error handling for sending requests
An error occured, when on user disabled the bot not by sending the /stop command, but by blocking the bot. Since the bot doesn't know, that the user doesn't want to engage in a conversation, this will cause the bot to send the message anyway. The message is not permitted to go through, so the Telegram API returns a 403 - Request Forbidden, which is mapped to a TelegramAPIException. This commit catches the most common exception types and logs the output.
Since finding the error was not possible, new messages are logged now.