Receiving voice messages is enabled
The user can send a voice message to the bot, which will react to the message and will download the voice message from the telegram servers. Only messages shorter than 7 seconds will be accepted and then downloaded. This is done in preparation for the voice commands, which will be limited in order to save API Calls/quota.