Move Abo settings over to general settings
In prvious versions /abo is used to configure one's abonnement. A user could choose to receive notifications about the courses INFB, MKIB and INFM and to faculty news. These settings are moved over to the new /settings command. This approach is more usable and intuitive. In addition to that, it is more responsive, since the user has the option to navigate the options.