Receiving and printing basic timetable information is now possible
Receiving the timetable information from the faculty servers is enabled
now. This is only done with an example, since the necessary information
to create the url are usually received from the database, which doesn't
have the capability to store such data for now.
The json is correctly interpreted and can be returned in a chatbot
fashion.