Uses vobis to analyze voice file
The voice file wasn't recognized correctly, if different rates (in hz)
were used. The desktop client uses 48000 hz and the mobile client, on
Android, uses 16000 hz. This causes problems, so the hz need to be
calculated by anaylzing the voice file (in ogg format). To do this, a
library is used.