Split up rive files to avoid complexity
Having everything in one giant file is kinda bad, so funny/snarky
answers are in a seperate file to better distinguis between serious
responses and funny ones. In addition, arrays used to be in the main
file but are now moved over to a seperate file in order to better
organize the project.

Matching agains professors is improved.